+++ packages/eslint-plugin/CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Changed
+
+- Removed `jsx-a11y/autocomplete-valid` rule for `eslint-plugin`([264](https://github.com/Shopify/web-configs/pull/264))
+- Add `react-require-autocomplete` lint rule to `@shopify/eslint-plugin` https://github.com/Shopify/web-configs/pull/251
 
 ## 40.2.3 - 2021-05-10